fixed translation and formatting issues in invoices #1670
