+++ src/Invoice/Renderer/TwigRenderer.php
@@ -11,22 +11,10 @@ namespace App\Invoice\Renderer;
 
 use App\Entity\InvoiceDocument;
 use App\Invoice\InvoiceModel;
-use App\Invoice\RendererInterface;
 use Symfony\Component\HttpFoundation\Response;
-use Twig\Environment;
 
-final class TwigRenderer implements RendererInterface
+final class TwigRenderer extends AbstractTwigRenderer
 {
-    /**
-     * @var Environment
-     */
-    private $twig;
-
-    public function __construct(Environment $twig)
-    {
-        $this->twig = $twig;
-    }
-
     public function supports(InvoiceDocument $document): bool
     {
         return stripos($document->getFilename(), '.html.twig') !== false;
@@ -34,9 +22,7 @@ final class TwigRenderer implements RendererInterface
 
     public function render(InvoiceDocument $document, InvoiceModel $model): Response
     {
-        $content = $this->twig->render('@invoice/' . basename($document->getFilename()), [
-            'model' => $model
-        ]);
+        $content = $this->renderTwigTemplate($document, $model);
 
         $response = new Response();
         $response->setContent($content);
