+++ templates/invoice/renderer/default.html.twig
@@ -1,48 +1,46 @@
 {% extends 'invoice/layout.html.twig' %}
-{% set fallback = app.request is not null ? app.request.locale : 'en' %}
-{% set language = model.template.language|default(fallback) %}
-{% set isDecimal = model.template.decimalDuration|default(false) %}
-{% set currency = model.currency %}
 
 {% block invoice %}
+{% set isDecimal = model.template.decimalDuration|default(false) %}
+{% set currency = model.currency %}
 <div class="row">
     <div class="col-xs-12">
         <h2 class="page-header">
             <span contenteditable="true">{{ model.template.title }}</span>
-            <small class="pull-right">{{ 'label.date'|trans({}, 'messages', language) }}: {{ model.invoiceDate|date_short }}</small>
+            <small class="pull-right">{{ 'label.date'|trans }}: {{ model.invoiceDate|date_short }}</small>
         </h2>
     </div>
 </div>
 
 <div class="row">
     <div class="col-sm-5">
-        {{ 'invoice.to'|trans({}, 'messages', language) }}
+        {{ 'invoice.to'|trans }}
         <address contenteditable="true">
             <strong>{{ model.customer.company|default(model.customer.name) }}</strong><br>
             {{ model.customer.address|nl2br }}
             {% if model.customer.vatId is not empty %}
                 <br>
-                {{ 'label.vat_id'|trans({}, 'messages', language) }}: {{ model.customer.vatId }}
+                {{ 'label.vat_id'|trans }}: {{ model.customer.vatId }}
             {% endif %}
             {% if model.customer.number is not empty %}
                 <br>
-                {{ 'label.number'|trans({}, 'messages', language) }}: {{ model.customer.number }}
+                {{ 'label.number'|trans }}: {{ model.customer.number }}
             {% endif %}
             {% if model.query.project is not empty and model.query.project.orderNumber is not empty %}
                 <br>
-                {{ 'label.orderNumber'|trans({}, 'messages', language) }}: {{ model.query.project.orderNumber }}
+                {{ 'label.orderNumber'|trans }}: {{ model.query.project.orderNumber }}
             {% endif %}
         </address>
     </div>
     <div class="col-sm-2"></div>
     <div class="col-sm-5">
-        {{ 'invoice.from'|trans({}, 'messages', language) }}
+        {{ 'invoice.from'|trans }}
         <address contenteditable="true">
             <strong>{{ model.template.company }}</strong><br>
             {{ model.template.address|trim|nl2br }}
             {% if model.template.vatId is not empty %}
             <br>
-            {{ 'label.vat_id'|trans({}, 'messages', language) }}:
+            {{ 'label.vat_id'|trans }}:
             {{ model.template.vatId }}
             {% endif %}
         </address>
@@ -52,11 +50,11 @@
 <div class="row">
     <div class="col-sm-5">
         <p contenteditable="true">
-            <strong>{{ 'invoice.number'|trans({}, 'messages', language) }}:</strong>
+            <strong>{{ 'invoice.number'|trans }}:</strong>
             {{ model.invoiceNumber }}
             
             <br>
-            <strong>{{ 'invoice.due_days'|trans({}, 'messages', language) }}:</strong>
+            <strong>{{ 'invoice.due_days'|trans }}:</strong>
             {{ model.dueDate|date_short }}
         </p>
     </div>
@@ -68,11 +66,11 @@
         <table class="table">
             <thead>
                 <tr>
-                    <th>{{ 'label.date'|trans({}, 'messages', language) }}</th>
-                    <th>{{ 'label.description'|trans({}, 'messages', language) }}</th>
-                    <th class="text-right">{{ 'label.unit_price'|trans({}, 'messages', language) }}</th>
-                    <th class="text-right">{{ 'label.amount'|trans({}, 'messages', language) }}</th>
-                    <th class="text-right">{{ 'label.total_rate'|trans({}, 'messages', language) }}</th>
+                    <th>{{ 'label.date'|trans }}</th>
+                    <th>{{ 'label.description'|trans }}</th>
+                    <th class="text-right">{{ 'label.unit_price'|trans }}</th>
+                    <th class="text-right">{{ 'label.amount'|trans }}</th>
+                    <th class="text-right">{{ 'label.total_rate'|trans }}</th>
                 </tr>
             </thead>
             <tbody>
@@ -102,19 +100,19 @@
             <tfoot>
                 <tr>
                     <td colspan="4" class="text-right">
-                        {{ 'invoice.subtotal'|trans({}, 'messages', language) }}
+                        {{ 'invoice.subtotal'|trans }}
                     </td>
                     <td class="text-right">{{ model.calculator.subtotal|money(currency) }}</td>
                 </tr>
                 <tr>
                     <td colspan="4" class="text-right">
-                        {{ 'invoice.tax'|trans({}, 'messages', language) }} ({{ model.calculator.vat }}%)
+                        {{ 'invoice.tax'|trans }} ({{ model.calculator.vat }}%)
                     </td>
                     <td class="text-right">{{ model.calculator.tax|money(currency) }}</td>
                 </tr>
                 <tr>
                     <td colspan="4" class="text-right text-nowrap">
-                        <strong>{{ 'invoice.total'|trans({}, 'messages', language) }}</strong>
+                        <strong>{{ 'invoice.total'|trans }}</strong>
                     </td>
                     <td class="text-right">
                         <strong>{{ model.calculator.total|money(currency) }}</strong>
@@ -137,12 +135,11 @@
 
 <footer class="footer">
     <p>
-        <strong>{{ 'label.address'|trans({}, 'messages', language) }}</strong>: {{ model.template.company }} &ndash; {{ model.template.address|replace({"\n": ' &ndash; ', "\r\n": ' &ndash; ', "\r": ' &ndash; '})|raw }}
+        <strong>{{ 'label.address'|trans }}</strong>: {{ model.template.company }} &ndash; {{ model.template.address|replace({"\n": ' &ndash; ', "\r\n": ' &ndash; ', "\r": ' &ndash; '})|raw }}
         <br>
-        <strong>{{ 'label.invoice_bank_account'|trans({}, 'messages', language) }}</strong>: {{ model.template.paymentDetails|replace({"\n": ' &ndash; ', "\r\n": ' &ndash; ', "\r": ' &ndash; '})|raw }}
+        <strong>{{ 'label.invoice_bank_account'|trans }}</strong>: {{ model.template.paymentDetails|replace({"\n": ' &ndash; ', "\r\n": ' &ndash; ', "\r": ' &ndash; '})|raw }}
         <br>
-        <strong>{{ 'label.contact'|trans({}, 'messages', language) }}</strong>: {{ model.template.contact|replace({"\n": ' &ndash; ', "\r\n": ' &ndash; ', "\r": ' &ndash; '})|raw }}
+        <strong>{{ 'label.contact'|trans }}</strong>: {{ model.template.contact|replace({"\n": ' &ndash; ', "\r\n": ' &ndash; ', "\r": ' &ndash; '})|raw }}
     </p>
 </footer>
-
 {% endblock %}
