+++ templates/widget/widget-paginatedworkingtimechart.html.twig
@@ -53,7 +53,7 @@
                     <div class="col-sm-3 col-xs-6">
                         <div class="description-block border-right">
                             <h5 class="description-header">{{ data.month|duration }}</h5>
-                            <span class="description-text">{{ 'stats.workingTimeMonth'|trans({'%month%': data.begin|month_name|trans, '%year%': data.begin|date_format('Y')}) }}</span>
+                            <span class="description-text">{{ 'stats.workingTimeMonth'|trans({'%month%': data.begin|month_name, '%year%': data.begin|date_format('Y')}) }}</span>
                         </div>
                     </div>
                     <div class="col-sm-3 col-xs-6">
