+++ translations/messages.es.xlf
@@ -914,58 +914,6 @@
             </trans-unit>
 
             <!--
-              Month names
-            -->
-            <trans-unit id="month.1">
-                <source>month.1</source>
-                <target>Enero</target>
-            </trans-unit>
-            <trans-unit id="month.2">
-                <source>month.2</source>
-                <target>Febrero</target>
-            </trans-unit>
-            <trans-unit id="month.3">
-                <source>month.3</source>
-                <target>Marzo</target>
-            </trans-unit>
-            <trans-unit id="month.4">
-                <source>month.4</source>
-                <target>Abril</target>
-            </trans-unit>
-            <trans-unit id="month.5">
-                <source>month.5</source>
-                <target>Mayo</target>
-            </trans-unit>
-            <trans-unit id="month.6">
-                <source>month.6</source>
-                <target>Junio</target>
-            </trans-unit>
-            <trans-unit id="month.7">
-                <source>month.7</source>
-                <target>Julio</target>
-            </trans-unit>
-            <trans-unit id="month.8">
-                <source>month.8</source>
-                <target>Agosto</target>
-            </trans-unit>
-            <trans-unit id="month.9">
-                <source>month.9</source>
-                <target>Septiembre</target>
-            </trans-unit>
-            <trans-unit id="month.10">
-                <source>month.10</source>
-                <target>Octubre</target>
-            </trans-unit>
-            <trans-unit id="month.11">
-                <source>month.11</source>
-                <target>Noviembre</target>
-            </trans-unit>
-            <trans-unit id="month.12">
-                <source>month.12</source>
-                <target>Diciembre</target>
-            </trans-unit>
-
-            <!--
              Navbar - recent entries and activities
             -->
             <trans-unit id="active.entries">
