+++ translations/messages.nl.xlf
@@ -914,58 +914,6 @@
             </trans-unit>
 
             <!--
-              Month names
-            -->
-            <trans-unit id="month.1">
-                <source>month.1</source>
-                <target>Januari</target>
-            </trans-unit>
-            <trans-unit id="month.2">
-                <source>month.2</source>
-                <target>Februari</target>
-            </trans-unit>
-            <trans-unit id="month.3">
-                <source>month.3</source>
-                <target>Maart</target>
-            </trans-unit>
-            <trans-unit id="month.4">
-                <source>month.4</source>
-                <target>April</target>
-            </trans-unit>
-            <trans-unit id="month.5">
-                <source>month.5</source>
-                <target>Mei</target>
-            </trans-unit>
-            <trans-unit id="month.6">
-                <source>month.6</source>
-                <target>Juni</target>
-            </trans-unit>
-            <trans-unit id="month.7">
-                <source>month.7</source>
-                <target>Juli</target>
-            </trans-unit>
-            <trans-unit id="month.8">
-                <source>month.8</source>
-                <target>Augustus</target>
-            </trans-unit>
-            <trans-unit id="month.9">
-                <source>month.9</source>
-                <target>September</target>
-            </trans-unit>
-            <trans-unit id="month.10">
-                <source>month.10</source>
-                <target>Oktober</target>
-            </trans-unit>
-            <trans-unit id="month.11">
-                <source>month.11</source>
-                <target>November</target>
-            </trans-unit>
-            <trans-unit id="month.12">
-                <source>month.12</source>
-                <target>December</target>
-            </trans-unit>
-
-            <!--
              Navbar - recent entries and activities
             -->
             <trans-unit id="active.entries">
