+++ translations/messages.ro.xlf
@@ -907,54 +907,7 @@
         <source>export.date_copyright</source>
         <target>Creat în %date% cu %kimai%</target>
       </trans-unit>
-      <trans-unit id="TzrTkzA" resname="month.1">
-        <source>month.1</source>
-        <target>Ianuarie</target>
-      </trans-unit>
-      <trans-unit id="Ybbwj3q" resname="month.2">
-        <source>month.2</source>
-        <target>Februarie</target>
-      </trans-unit>
-      <trans-unit id="ZUWOMta" resname="month.3">
-        <source>month.3</source>
-        <target>Martie</target>
-      </trans-unit>
-      <trans-unit id="r7dXujQ" resname="month.4">
-        <source>month.4</source>
-        <target>Aprilie</target>
-      </trans-unit>
-      <trans-unit id="5U8dGjy" resname="month.5">
-        <source>month.5</source>
-        <target>Mai</target>
-      </trans-unit>
-      <trans-unit id="S__KfXn" resname="month.6">
-        <source>month.6</source>
-        <target>Iunie</target>
-      </trans-unit>
-      <trans-unit id="c6TKQ2p" resname="month.7">
-        <source>month.7</source>
-        <target>Iulie</target>
-      </trans-unit>
-      <trans-unit id="js3tNpC" resname="month.8">
-        <source>month.8</source>
-        <target>August</target>
-      </trans-unit>
-      <trans-unit id="2thDj5n" resname="month.9">
-        <source>month.9</source>
-        <target>Septembrie</target>
-      </trans-unit>
-      <trans-unit id="HshzwEV" resname="month.10">
-        <source>month.10</source>
-        <target>Octombrie</target>
-      </trans-unit>
-      <trans-unit id="Qi4.VFn" resname="month.11">
-        <source>month.11</source>
-        <target>Noiembrie</target>
-      </trans-unit>
-      <trans-unit id="LXtp3N0" resname="month.12">
-        <source>month.12</source>
-        <target>Decembrie</target>
-      </trans-unit>
+
       <trans-unit id="OxdYMR3" resname="active.entries">
         <source>active.entries</source>
         <target>Înregistrările tale active</target>
